// src/app/model/personal-info.model.ts
export interface PersonalInfo {
  fullName: string;
  fullNameEnglish: string;
  city: string;
  street: string;
  building: string;
  floor: number;
  email: string;
  landline: number;
  parentsMobile: number;
  personalMobile: number;
  birthDate: Date;
  placeOfBirth: string;
  nationalityAtBirth: string;
  nationalityNow: string;
  gender: string;
  socialStatus: string;
  employmentStatus: string;
  companyName: string;
  companyAddress: string;
  lastSalary: string;
}